# src/Cifra.Api/V1/Models/Class/Responses/GetClassResponse.cs
namespace Cifra.Api.V1.Models.Class.Responses
{
    public class GetClassResponse
    {
        /// <summary>
        /// The class
        /// </summary>
        public Class RetrievedClass { get; set; }
    }
}